pyproject.toml
[project]
name = "novaenx_frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autopep8>=2.3.2",
    "django>=5.2.6",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "jupyter>=1.1.1",
    "notebook>=7.4.7",
    "pytz>=2025.2",
    "requests>=2.32.5",
]
